Disable chpasswd step
For some reason, the chpasswd --root command, outside the arch-chroot
context is not working as it was last week.

I've tried running an older version of the script and custom package
and the result was the same. Probably it is due to some breaking change
or a bug introduced in the latest Arch Linux's packages versions as we
always run pacman -Sy to make sure we're installing updated packages.

The error message was:
chpasswd: (user lemuel) pam_chauthtok() failed, error:
Module is unknown

As an alternative, we can set up the user's password by running chpasswd
inside an arch-chroot context without the --root parameter.

Links:

- https://www.linuxquestions.org/questions/slackware-14/error-creating-lxc-container-on-current-4175677752/
- https://www.reddit.com/r/voidlinux/comments/l2nwhk/error_during_installation_it_wont_create_an_user/
- http://www.huge-man-linux.net/man8/chpasswd.html
- https://docs.jdcloud.com/en/virtual-machines/ssh-login-report-module-is-unknown